AudioTrack: When paused, return cached playback position
An offload output can be re-used between two audio tracks having
the same configuration. A timestamp query for a paused track
while the other is running would return an incorrect time.
To fix this, cache the playback position on a pause() and return
this time when requested until the track is resumed.

Change-Id: I98e969abf49f8d3a9793ed1c065e21664d7553f7
CRs-Fixed: 612943