media: Match quality definitions with what the framework expects
These ids are used internally in order to parse the media codecs
quality definitions and pass them along to the frameworks/base component.

These ids must match the ones present in f/b/media/java/android/media/CamcorderProfile.java

If the hardware is a qualcomm, the match if OK. Otherwise though, they are
off-by-one which causes crashes in the camera app when selecting a low
resolution.

Change-Id: I4e83011c54114265b422e9430aa8612574f04e4a