audioflinger: Throttle output if no active tracks
Throttle output for the duration from when there are
no active tracks to the time output is put to
standby due to inactivity. This is done to save some
power during this duration. A typical usecase affected
by this optimization is key press on the dialer
keypad or keyboard with a delay of a few hundred
milliseconds between keys.

CRs-Fixed: 2088743
Change-Id: I5b73c18a7b7767568e2461c72d8d7f86cafc9b7e